seperate views ipc
