+++ src/shared/ipc/views.ipc.ts
@@ -1,215 +1,7 @@
-import { IPCBaseMessage } from './base.ipc'
-import { IWidget } from '../interfaces/widgets.interface';
-
-// List of tables
-export const IPC_CHANNEL_LIST_OF_TABLES: string = "channel_list_of_tables";
-
-export interface IIPCListOfTablesRequestMessage { }
-
-export interface IIPCListOfTablesResponseMessage {
-    valid: boolean;
-    error?: string;
-    tables?: string[]
-}
-
-export class IPCListOfTablesResponseMessage extends IPCBaseMessage<IIPCListOfTablesResponseMessage> { }
-
-export class IPCListOfTablesRequestMessage extends IPCBaseMessage<IIPCListOfTablesRequestMessage> { }
-
-
-// Table info
-export const IPC_CHANNEL_TABLE_INFO: string = "channel_table_info";
-
-export interface IIPCTableInfoRequestMessage {
-    table: string;
-}
-
-export interface IIPCTableInfoColumn {
-    name: string;
-    default: string;
-    nullable: boolean;
-    type: string;
-    length: number;
-    key: string;
-    extra: string;
-}
-
-export interface IIPCTableInfoResponseMessage {
-    valid: boolean;
-    error?: string;
-    columns?: IIPCTableInfoColumn[]
-}
-
-export class IPCTableInfoResponseMessage extends IPCBaseMessage<IIPCTableInfoResponseMessage> { }
-
-export class IPCTableInfoRequestMessage extends IPCBaseMessage<IIPCTableInfoRequestMessage> { }
-
-// Table read data
-export const IPC_CHANNEL_READ_DATA: string = "channel_read_data";
-
-export interface IIPCReadDataRequestMessage {
-    table: string;
-    limit: IIPCReadDataLimit;
-    columns: string[];
-    search?: IIPCReadDataSearch;
-    where?: IIPCReadDataWhere[]
-}
-
-export interface IIPCReadDataLimit {
-    limit: number;
-    offset: number;
-}
-
-export interface IIPCReadDataSearch {
-    columns: string[],
-    text: string
-}
-
-export interface IIPCReadDataWhere {
-    column: string;
-    opr: IIPCReadDataWhereOpr;
-    value: any;
-    or: boolean;
-}
-
-export enum IIPCReadDataWhereOpr {
-    EQ = "=",
-    GT = ">",
-    LT = "<"
-}
-
-export interface IIPCReadDataResponseMessage {
-    valid: boolean;
-    error?: string;
-    data?: {
-        [key: string]: any
-    }[],
-    count?: number;
-}
-
-export class IPCReadDataResponseMessage extends IPCBaseMessage<IIPCReadDataResponseMessage> { }
-
-export class IPCReadDataRequestMessage extends IPCBaseMessage<IIPCReadDataRequestMessage> { }
-
-
-// Table update data
-export const IPC_CHANNEL_UPDATE_DATA: string = "channel_update_data";
-
-export interface IIPCUpdateDataRequestMessage {
-    table: string;
-    update: IIPCUpdateDataUpdate;
-    where?: IIPCUpdateDataWhere[]
-}
-
-export interface IIPCUpdateDataUpdate {
-    [key: string]: any;
-}
-
-export interface IIPCUpdateDataWhere {
-    column: string;
-    opr: IIPCUpdateDataWhereOpr;
-    value: any;
-    or: boolean;
-}
-
-export enum IIPCUpdateDataWhereOpr {
-    EQ = "=",
-    GT = ">",
-    LT = "<"
-}
-
-export interface IIPCUpdateDataResponseMessage {
-    valid: boolean;
-    error?: string;
-}
-
-export class IPCUpdateDataResponseMessage extends IPCBaseMessage<IIPCUpdateDataResponseMessage> { }
-
-export class IPCUpdateDataRequestMessage extends IPCBaseMessage<IIPCUpdateDataRequestMessage> { }
-
-
-// Table insert data
-export const IPC_CHANNEL_INSERT_DATA: string = "channel_insert_data";
-
-export interface IIPCInsertDataRequestMessage {
-    table: string;
-    data: IIPCInsertData;
-}
-
-export interface IIPCInsertData {
-    [key: string]: any;
-}
-
-export interface IIPCInsertDataResponseMessage {
-    valid: boolean;
-    error?: string;
-}
-
-export class IPCInsertDataResponseMessage extends IPCBaseMessage<IIPCInsertDataResponseMessage> { }
-
-export class IPCInsertDataRequestMessage extends IPCBaseMessage<IIPCInsertDataRequestMessage> { }
-
-
-// Table delete data
-export const IPC_CHANNEL_DELETE_DATA: string = "channel_delete_data";
-
-export interface IIPCDeleteDataRequestMessage {
-    table: string;
-    where?: IIPCDeleteDataWhere[]
-}
-
-export interface IIPCDeleteDataWhere {
-    column: string;
-    opr: IIPCDeleteDataWhereOpr;
-    value: any;
-    or: boolean;
-}
-
-export enum IIPCDeleteDataWhereOpr {
-    EQ = "=",
-    GT = ">",
-    LT = "<"
-}
-
-export interface IIPCDeleteDataResponseMessage {
-    valid: boolean;
-    error?: string;
-}
-
-export class IPCDeleteDataResponseMessage extends IPCBaseMessage<IIPCDeleteDataResponseMessage> { }
-
-export class IPCDeleteDataRequestMessage extends IPCBaseMessage<IIPCDeleteDataRequestMessage> { }
-
-// Table read widget data
-export const IPC_CHANNEL_READ_WIDGET_DATA: string = "channel_read_widget_data";
-
-/**
- * Interface that represent a request message sent to IPC
- * The structure is same like IWidget, we just extends it
- */
-export interface IIPCReadWidgetDataRequestMessage extends IWidget {
-    // Table is missing in IWidget, we have to provide it
-    table: string;
- }
-
-/**
- * Interface that represent a response message 
- */
-export interface IIPCReadWidgetDataResponseMessage {
-    // Is the result valid
-    valid: boolean;
-    // In case of error, the error message
-    error?: string;
-    // The result in case that all works
-    data?: number;
-}
-
-/**
- * The request message class, implements IIPCReadWidgetDataResponseMessage
- */
-export class IPCReadWidgetDataResponseMessage extends IPCBaseMessage<IIPCReadWidgetDataResponseMessage> { }
-
-/**
- * The response message class, implements IIPCReadWidgetDataRequestMessage
- */
-export class IPCReadWidgetDataRequestMessage extends IPCBaseMessage<IIPCReadWidgetDataRequestMessage> { }
+export * from './views/tables-list.ipc';
+export * from './views/table-info.ipc';
+export * from './views/read-data.ipc';
+export * from './views/insert-data.ipc';
+export * from './views/update-data.ipc';
+export * from './views/delete-data.ipc';
+export * from './views/read-widget-data.ipc';
