+++ src/shared/ipc/views/delete-data.ipc.ts
@@ -1,0 +1,30 @@
+import { IPCBaseMessage } from "../base.ipc";
+
+export const IPC_CHANNEL_DELETE_DATA: string = "channel_delete_data";
+
+export interface IIPCDeleteDataRequestMessage {
+    table: string;
+    where?: IIPCDeleteDataWhere[]
+}
+
+export interface IIPCDeleteDataWhere {
+    column: string;
+    opr: IIPCDeleteDataWhereOpr;
+    value: any;
+    or: boolean;
+}
+
+export enum IIPCDeleteDataWhereOpr {
+    EQ = "=",
+    GT = ">",
+    LT = "<"
+}
+
+export interface IIPCDeleteDataResponseMessage {
+    valid: boolean;
+    error?: string;
+}
+
+export class IPCDeleteDataResponseMessage extends IPCBaseMessage<IIPCDeleteDataResponseMessage> { }
+
+export class IPCDeleteDataRequestMessage extends IPCBaseMessage<IIPCDeleteDataRequestMessage> { }
